// src/commentRepository/observers/getComments/enums.ts
export enum EnumCommentActions {
  OnCommentCreated = 'onCommentCreated',
  OnCommentUpdated = 'onCommentUpdated',
  OnCommentDeleted = 'onCommentDeleted',
  OnCommentFlagged = 'onCommentFlagged',
  OnCommentUnflagged = 'onCommentUnflagged',
  OnCommentReactionAdded = 'onCommentReactionAdded',
  OnCommentReactionRemoved = 'onCommentReactionRemoved',
}